+++ pkgs/google_cloud_pubsub/lib/google_cloud_pubsub.dart
@@ -22,5 +22,6 @@ export 'package:grpc/grpc.dart'
 
 export 'src/client.dart' show PubSub;
 export 'src/message.dart' show Message, ReceivedMessage;
+export 'src/retry.dart' show RetrySettings;
 export 'src/subscription.dart' show Subscription;
 export 'src/topic.dart' show Topic;